+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.6.0] - May 11, 2026
+
+### Changed
+
+- Use the Goodreads autocomplete API instead of scraping HTML for search results
+
+### Fixed
+
+- Fix no search results - [(ISSUE)](https://github.com/nesaku/BiblioReads/issues/40)
+
 ## [3.5.4] - Jan 24, 2026
 
 ### Changed